test(delta): move MinIO S3 test to tests/ to fix coverage linker OOM
The #[ignore] MinIO test lived in the delta module's #[cfg(test)], which linked
testcontainers/bollard into the connectors lib test binary. Under the coverage
job (cargo llvm-cov --all-features, instrumented) that bloated binary overflowed
the runner's linker (ld SIGBUS -> "could not compile ... (lib test)"). The crash
is non-deterministic across crates (db or connectors) since it's parallel-link
memory pressure; coverage was green before this test was added.

Every other testcontainers test already lives in tests/. Move it to
tests/delta_s3_integration.rs (its own small integration binary, gated on
delta-lake-s3, #[ignore]) and revert the read_regions split it required. No
behavior change; the lib test binary no longer links testcontainers.
